Adding new rigs and fix spine stretch to
In the steps to make the rig more universal for retargeting across engines, it's neede da few overhauls. No doubt this will break some compatiblity and may frustrate some users. But overall the net gain is worth it.

Changes in rig name make it happy to use with Unreal. The Fix of the spine 2 "stretch_to" constraint fixes the hunchback issue the former rig had. And new metraigs have been added for each Quaternius base character type to allow for easily working with his base meshes and clothing.